requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.15.0
numpy==1.24.3
pandas==2.1.3
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
locust==2.17.0
jupyter==1.0.0
ipykernel==6.27.1
opencv-python==4.8.1.78
albumentations==1.3.1
tqdm==4.66.1 